requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.5
django-huey==1.1.1
django-ninja==0.18.0
gevent==21.12.0
greenlet==1.1.2
huey==2.4.3
packaging==21.3
pydantic==1.9.1
pyparsing==3.0.9
python-environ==0.4.54
redis==4.2.2
sqlparse==0.4.2
typing_extensions==4.2.0
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
